Added home page to pattern lab
Enhancement to nav bar making it more obvious to users how to return to
the default state of patternlab, rather than relying on page reload or
history.

this is not updating the history stack yet, I was hoping Dave could help
me with that if this pull request is approved in the first place.  I
think that net, this is a usability win.  When I first used Pattern Lab
and clicked on a pattern, I struggled with how to return.